Move recipe parser out of operators submodule
