+++ tests/test_recipe_parser.py
@@ -19,7 +19,7 @@ import tempfile
 from uuid import uuid4
 
 import CSET.operators.RECIPES as RECIPES
-import CSET.operators._internal as internal
+import CSET._recipe_parsing as recipe_parsing
 
 
 logging.basicConfig(level=logging.DEBUG)
@@ -27,13 +27,13 @@ logging.basicConfig(level=logging.DEBUG)
 
 def test_get_operator():
     """Get operator and test exceptions"""
-    read_operator = internal.get_operator("read.read_cubes")
+    read_operator = recipe_parsing.get_operator("read.read_cubes")
     assert callable(read_operator)
 
     # Test exception for non-existent operators.
     exception_occurred = False
     try:
-        internal.get_operator("non-existent.operator")
+        recipe_parsing.get_operator("non-existent.operator")
     except ValueError:
         exception_occurred = True
     assert exception_occurred
@@ -41,7 +41,7 @@ def test_get_operator():
     # Test exception if wrong type provided.
     exception_occurred = False
     try:
-        internal.get_operator(["Not", b"a", "string", 1])
+        recipe_parsing.get_operator(["Not", b"a", "string", 1])
     except ValueError:
         exception_occurred = True
     assert exception_occurred
@@ -49,7 +49,7 @@ def test_get_operator():
     # Test exception if operator isn't a function.
     exception_occurred = False
     try:
-        internal.get_operator("RECIPES.extract_instant_air_temp")
+        recipe_parsing.get_operator("RECIPES.extract_instant_air_temp")
     except ValueError:
         exception_occurred = True
     assert exception_occurred
@@ -62,7 +62,9 @@ def test_execute_recipe():
     # Test exception for non-existent file.
     exception_happened = False
     try:
-        internal.execute_recipe(Path("/non-existent/path"), os.devnull, os.devnull)
+        recipe_parsing.execute_recipe(
+            Path("/non-existent/path"), os.devnull, os.devnull
+        )
     except FileNotFoundError:
         exception_happened = True
     assert exception_happened
@@ -70,7 +72,7 @@ def test_execute_recipe():
     # Test exception for incorrect type.
     exception_happened = False
     try:
-        internal.execute_recipe(True, os.devnull, os.devnull)
+        recipe_parsing.execute_recipe(True, os.devnull, os.devnull)
     except TypeError:
         exception_happened = True
     assert exception_happened
@@ -78,7 +80,7 @@ def test_execute_recipe():
     # Test exception for invalid YAML.
     exception_happened = False
     try:
-        internal.execute_recipe(
+        recipe_parsing.execute_recipe(
             '"Inside quotes" outside of quotes', os.devnull, os.devnull
         )
     except ValueError:
@@ -88,7 +90,7 @@ def test_execute_recipe():
     # Test exception for valid YAML but invalid recipe.
     exception_happened = False
     try:
-        internal.execute_recipe("a: 1", os.devnull, os.devnull)
+        recipe_parsing.execute_recipe("a: 1", os.devnull, os.devnull)
     except ValueError:
         exception_happened = True
     assert exception_happened
@@ -96,7 +98,7 @@ def test_execute_recipe():
     # Test exception for blank recipe.
     exception_happened = False
     try:
-        internal.execute_recipe("", os.devnull, os.devnull)
+        recipe_parsing.execute_recipe("", os.devnull, os.devnull)
     except ValueError:
         exception_happened = True
     assert exception_happened
@@ -104,7 +106,7 @@ def test_execute_recipe():
     # Test exception for recipe without any steps.
     exception_happened = False
     try:
-        internal.execute_recipe("steps: []", os.devnull, os.devnull)
+        recipe_parsing.execute_recipe("steps: []", os.devnull, os.devnull)
     except ValueError:
         exception_happened = True
     assert exception_happened
@@ -112,7 +114,7 @@ def test_execute_recipe():
     # Test exception for recipe that parses to a non-dict.
     exception_happened = False
     try:
-        internal.execute_recipe("[]", os.devnull, os.devnull)
+        recipe_parsing.execute_recipe("[]", os.devnull, os.devnull)
     except ValueError:
         exception_happened = True
     assert exception_happened
@@ -120,12 +122,12 @@ def test_execute_recipe():
     # Test happy case (this is really an integration test).
     output_file = Path(f"{tempfile.gettempdir()}/{uuid4()}.nc")
     recipe_file = RECIPES.extract_instant_air_temp
-    internal.execute_recipe(recipe_file, input_file, output_file)
+    recipe_parsing.execute_recipe(recipe_file, input_file, output_file)
     output_file.unlink()
 
     # Test weird edge cases. Also tests paths not being pathlib Paths, and
     # directly passing in a stream for a recipe file.
     output_file = f"{tempfile.gettempdir()}/{uuid4()}.nc"
     with open("tests/test_data/noop_recipe.yaml", "rb") as recipe:
-        internal.execute_recipe(recipe, str(input_file), output_file)
+        recipe_parsing.execute_recipe(recipe, str(input_file), output_file)
     # The output_file doesn't actually get written, so doesn't need removing.
